legacy/requirements.txt
# ============================================================
# E-commerce Audit System — Python dependencies
# Environment: CUDA 12.4, Python 3.10 (conda)
# ============================================================

# ---- PyTorch (installed via conda/pip with CUDA 12.4 index) ----
# torch>=2.4.0        # installed separately with correct CUDA wheel
# torchvision>=0.19.0 # installed alongside torch
# torchaudio>=2.4.0   # installed alongside torch

# ---- Transformers ecosystem ----
transformers>=4.52.0
tokenizers>=0.21.0
accelerate>=1.2.0
peft>=0.14.0
datasets>=3.0.0
evaluate>=0.4.3
huggingface_hub>=0.27.0

# ---- VLM & multimodal ----
qwen-vl-utils>=0.0.8
Pillow>=10.4.0

# ---- RL / training infrastructure ----
# veRL / FIPO installed from source (see setup-env instructions)
deepspeed>=0.15.0
ray[default]>=2.20.0

# ---- Inference engine ----
vllm>=0.8.0

# ---- RAG ----
faiss-gpu-cu12>=1.9.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# ---- Data processing ----
pyarrow>=16.0.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0

# ---- Experiment tracking ----
wandb>=0.18.0
tensorboard>=2.17.0

# ---- API clients (Stage 0 distillation via DashScope) ----
openai>=1.50.0
dashscope>=1.20.0
httpx>=0.27.0

# ---- Dataset sources ----
modelscope>=1.18.0

# ---- Config management ----
hydra-core>=1.3.2
omegaconf>=2.3.0

# ---- Utilities ----
rich>=13.7.0
pytest>=8.0.0
